Refactor resolvers
